requirements.txt
langchain
langgraph
langchain-openai
langchain-community
faiss-cpu
pydantic
python-dotenv
